@@ -6,23 +6,39 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using ValoGun.Models.Player_Cards;
+using Microsoft.Maui.ApplicationModel;
+using Microsoft.Maui.Storage;
+using CommunityToolkit.Mvvm.Input;
 
 namespace ValoGun.ViewModels
 {
-	public class PlayerCardsViewModel : BaseViewModel
+	public partial class PlayerCardsViewModel : BaseViewModel
 	{
 
+		public string SearchText { get; set; }
 
-		public ObservableRangeCollection<Data> _PlayerCards { get; set; }
+		public static ObservableRangeCollection<Data> PlayerCards { get; set; }
 
 		public PlayerCardsViewModel()
 		{
-			_PlayerCards = [];
+			PlayerCards = [];
 
-			Thread DataThread = new Thread(async () => await LoadPlayerCards());
+			Thread DataThread = new(async () => await LoadPlayerCards());
 			DataThread.Start();
 		}
+		
+		public static Task<ObservableRangeCollection<Data>> Search(string SearchText = "")
+        {
+            if(string.IsNullOrEmpty(SearchText))
+            {
+                return null;
+            }
 
+            var SearchResults = new ObservableRangeCollection<Data>();
+            SearchResults.AddRange(PlayerCards.Where(x => x.displayName.Contains(SearchText, StringComparison.CurrentCultureIgnoreCase)));
+
+            return Task.FromResult(SearchResults);
+        }
 		private async Task LoadPlayerCards()
 		{
 			using var stream = await FileSystem.OpenAppPackageFileAsync("PlayerCards.json");
@@ -33,10 +49,10 @@ namespace ValoGun.ViewModels
 
 			foreach(var card in Cards.data)
 			{
-				await MainThread.InvokeOnMainThreadAsync(() => _PlayerCards.Add(card));
-				await MainThread.InvokeOnMainThreadAsync(() => OnPropertyChanged(nameof(_PlayerCards)));
+				await MainThread.InvokeOnMainThreadAsync(() => PlayerCards.Add(card));
+				await MainThread.InvokeOnMainThreadAsync(() => OnPropertyChanged(nameof(PlayerCards)));
 			}
-			await MainThread.InvokeOnMainThreadAsync(async () => { await Shell.Current.DisplayAlert("Finished", "Finished", "OK"); });
+			//await MainThread.InvokeOnMainThreadAsync(async () => await Shell.Current.DisplayAlert("Finished", "Finished", "OK"));
 
 		}
 	}
